2.5.3 - end of a hiatus.
+ added Mute Command
* setMutedRole and SetAdminRole can now use the arguments remove to
remove the role from the config.
* fixed DenyInvites talking about Custom commands
+ module ModMute Added.
* if the user is the server owner no matter what, they will now have the
max cc slots.
* Replaces userMentions in Edit and Delete logging with Plain text.
+ added toggles : ChannelLogging, ExtendEditLog, UseTimestamps,
slashCommands.
+ added Slash commands
/DealWithIt,/Disapprove,/Fite,/Gib,/Lenny,/Shrug,/TableFlip,/UnFlip
+ added command SlashCommands
~~* Nope No jackson here only Gson~~
~~* Saved Hyrule.~~

+++ src/main/java/Objects/UserTypeObject.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
  */
 public class UserTypeObject {
     String ID;
-    float xp = 0;
+    long xp = 0;
     long level = 0;
     String rewardRoleID = "";
     String gender = "Unknown";
@@ -66,7 +66,7 @@ public class UserTypeObject {
         level += 1;
     }
 
-    public void setXp(float xp) {
+    public void setXp(long xp) {
         this.xp = xp;
         calculateLevel();
     }
